feat(fluids): Extract newtownian.h from DC problem
 - Also trade out `int` for `StabilizationType` enum
